+++ packages/scene/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @glissade/scene
 
+## 0.41.1-pre.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @glissade/core@0.41.1-pre.0
+
 ## 0.41.0
 
 ### Patch Changes